Update ROS wrapper to version 2.0 (BETA VERSION)
Add support for the 2.0 SDK.

This commit also contains some adjustments to the API of the ROS
driver, including renaming of topics, some new topics, and renaming
of dynamic_reconfigure nodes, etc. See CHANGELOG.md for full list
of publicly visible changes in this release.

Internally, most of the changes are to work with API 2.0. There is also
some cleanup and refactoring. For example, handling of Settings and
Settings2D configuration is moved to a shared class
capture_settings_controller.h. This ensures that the settings and
settings_2d dynamic_reconfigure trees are consistent.

The depth image topic has been renamed from "depth/image_raw" to
"depth/image". Since this image contains depth as 32-bit floats, it
should be named "image" to follow the recommendation in
[REP-118: Depth Images](https://www.ros.org/reps/rep-0118.html).

test_zivid_camera.cpp now does output testing by capturing using the
SDK directly, and comparing the output from the SDK with the output
from ROS. Previously, testing was using hard-coded values for
MiscObjects.zdf. However, when testing ROS we can assume that the
output from SDK is good.