Remove buggy zivid_camera_with_settings.launch file
This launch file tried to set 3-frame HDR; but once in a while it
would not update all of the settings. It appears that one or more
of the dynparam nodes being created end up hanging forever (they
remain running in rosparam list). I am not sure why this is. For now
I am removing this sample .launch file. I have made issue 31 to for
reintroducing this sample.